+++ src/main/java/dude/command/AddDeadlineCommand.java
@@ -2,11 +2,11 @@ package dude.command;
 
 import java.time.LocalDateTime;
 
-import dude.Storage;
 import dude.note.NoteList;
 import dude.task.Deadline;
 import dude.task.TaskList;
 import dude.ui.Ui;
+import dude.util.Storage;
 
 /**
  * Represents a command that adds a Deadline task.
