+++ src/main/java/dude/command/DeleteNoteCommand.java
@@ -1,10 +1,10 @@
 package dude.command;
 
-import dude.Storage;
 import dude.note.Note;
 import dude.note.NoteList;
 import dude.task.TaskList;
 import dude.ui.Ui;
+import dude.util.Storage;
 
 /**
  * Represents a command that deletes note.
