+++ src/main/java/dude/command/ExitCommand.java
@@ -1,9 +1,9 @@
 package dude.command;
 
-import dude.Storage;
 import dude.note.NoteList;
 import dude.task.TaskList;
 import dude.ui.Ui;
+import dude.util.Storage;
 
 /**
  * Represents a command that exits the programme.
