+++ src/main/java/dude/command/UnknownCommand.java
@@ -1,9 +1,9 @@
 package dude.command;
 
-import dude.Storage;
 import dude.note.NoteList;
 import dude.task.TaskList;
 import dude.ui.Ui;
+import dude.util.Storage;
 
 /**
  * Represents a command that is not defined in Dude.
