+++ src/main/java/dude/util/Storage.java
@@ -1,4 +1,4 @@
-package dude;
+package dude.util;
 
 import java.io.File;
 import java.io.FileNotFoundException;
